Refactor page into smaller components
